Don't suggest to replace d2dwy(date, 3) with date.Year() (#1035)
* Don't suggest to replace d2dwy(date, 3) with date.Year()

When the input date to the Date2DWY method is in a week that spans two years, the Date2DWY method computes the output year as the year that has the most days. For example, the input date is 010114. This date is in a week that starts on Monday, December 29, 2013, and ends Sunday, January 4, 2014. The week has three days in 2013 and four days in 2014. So the output year is 2014.
See: https://learn.microsoft.com/en-us/dynamics365/business-central/dev-itpro/developer/methods-auto/system/system-date2dwy-method#remarks

* add comment

---------

Co-authored-by: Arthur van de Vondervoort <44637996+Arthurvdv@users.noreply.github.com>
Co-authored-by: Kilian Seizinger <k.seizinger@prisma-informatik.de>